+++ src/components/editor.rs
@@ -372,7 +372,10 @@ impl<'a> Component<Msg, TisqEvent> for Editor<'a> {
                 kind: KeyEventKind::Press,
                 modifiers,
                 ..
-            }) if !ch.is_alphabetic() || modifiers == KeyModifiers::NONE => {
+            }) if !ch.is_alphabetic()
+                || modifiers == KeyModifiers::NONE
+                || modifiers == KeyModifiers::SHIFT =>
+            {
                 // either a non-alphabetic char or an alphabetic char without modifiers to allow
                 // for global key bindings on alphabetic chars
                 self.perform(Cmd::Type(ch));