+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - `--version` flag to print version.
+- Support arrays.
 
 ## [v0.1.0] - 2023-10-08