+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - ReleaseDate
 
+## [v0.1.6] - 2023-10-14
+
 ### Added
 
 - Cycle navigation shortcut Ctrl+n.
@@ -71,7 +73,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Execution results view supporting only successful fetched results.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/strowk/tisq/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/strowk/tisq/compare/v0.1.6...HEAD
+[v0.1.6]: https://github.com/strowk/tisq/compare/v0.1.5...v0.1.6
 [v0.1.5]: https://github.com/strowk/tisq/compare/v0.1.4...v0.1.5
 [v0.1.4]: https://github.com/strowk/tisq/compare/v0.1.3...v0.1.4
 [v0.1.3]: https://github.com/strowk/tisq/compare/v0.1.2...v0.1.3